refactor: move BioCardListModal into GenericPagesComponents
