+++ src/components/GenericPagesComponents/BioCardListModal/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./BioModal";
+export type { BioData } from "./BioModal";
