+++ src/components/GenericPagesComponents/BioCardListModal/useBioModal.ts
@@ -6,10 +6,11 @@ import {
   MutableRefObject,
 } from "react";
 
-import errorReporter from "../../common-lib/error-reporter";
-import OakError from "../../errors/OakError";
-
 import { BioData, BioModalProps } from "./BioModal";
+
+import errorReporter from "@/common-lib/error-reporter";
+import OakError from "@/errors/OakError";
+
 
 const reportError = errorReporter("useBioModal");
